+++ NEWS.md
@@ -1,7 +1,15 @@
+# Version 1.0.7
+
+## Changes and additions
+
+* changes required by CRAN for saving files in examples
+* imp_raven() error message when not all selections have a sound file name column
+
 # Version 1.0.6
 
 ## Changes and additions
 
+* unique names to temporal .txt output files to allow parallelization in 'raven_batch_detec()'
 * new arguments in 'imp_raven()': 'warbler.format' for importing with warbleR selection table column names, 'only.spectro.view' for keeping measurements only for the spectrogram view and 'files' for importing specific selection files
 
 # Version 1.0.5